reorganize steam api responses
