+++ steamguard/src/api_responses/phone_ajax.rs
@@ -1,0 +1,11 @@
+use serde::Deserialize;
+
+#[derive(Debug, Clone, Deserialize)]
+#[allow(dead_code)]
+pub struct PhoneValidateResponse {
+	success: bool,
+	number: String,
+	is_valid: bool,
+	is_voip: bool,
+	is_fixed: bool,
+}
